requirements.txt
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
absl-py==0.9.0
astor==0.8.1
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mock==4.0.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dateutil==2.8.1
PyYAML==5.3
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
tornado==6.0.4
Werkzeug==0.16.0
